Move fields context to ts
